+++ .gitignore
@@ -31,3 +31,4 @@
 *.out
 *.app
 main
+.vscode
