+++ main.cpp
@@ -1,10 +1,22 @@
+
 #include <iostream>
+/*
+- Function is a not first class citizen (not passable, not returnable). Use Lambda
+
+*/
 
 using namespace std;
+int a = 4;
+
+void func(){
+  int a = 4;
+  cout << "value: " << ::a << endl;
+
+}
 
 int main() {
-    int a = 0;
-    a += 1;
-    cout << "hello world" << endl;
-    return 0;
+  int a = 3;
+  int *ptr = &a;
+  func();
 }
+
