+++ readme.md
@@ -1,0 +1,6 @@
+# Notes
+
+
+## General
+- Function is a not first class citizen (not passable, not returnable). Use Lambda
+- `::` is used to access a global variable
